only documentation left

+++ src/main/java/com/example/service/MessageService.java
@@ -2,7 +2,6 @@ package com.example.service;
 
 import org.springframework.stereotype.Service;
 
-import com.example.entity.Account;
 import com.example.entity.Message;
 import com.example.repository.MessageRepository;
 
@@ -10,7 +9,6 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-
 
 @Service
 public class MessageService {
@@ -22,9 +20,10 @@ public class MessageService {
     public MessageService(MessageRepository messageRepository) {
         this.messageRepository = messageRepository;
     }
+
     public Message addMessage(Message message) {
         String messageContent = message.getMessageText();
-        
+
         if (messageContent.length() == 0 || messageContent.length() >= 255) {
             return null;
         }
@@ -32,7 +31,7 @@ public class MessageService {
         int posted_by = message.getPostedBy();
 
         if (!this.accountService.accountExists(posted_by)) {
-            throw new IllegalArgumentException();
+            return null;
         }
         return this.messageRepository.save(message);
     }
@@ -47,28 +46,26 @@ public class MessageService {
 
     public boolean updateMessage(int message_id, Message message) {
         String messageContent = message.getMessageText();
-        
+
         if (messageContent.length() == 0 || messageContent.length() >= 255) {
             return false;
         }
-
-        // int posted_by = message.getPostedBy();
-
-        // if (!this.accountService.accountExists(posted_by)) {
-        //     return false;
-        // }
         Optional<Message> optionalMessage = this.getMessageByID(message_id);
-        if (optionalMessage.isPresent()) {
-            Message updatedMessage = message;
-            updatedMessage.setMessageId(message_id);
-            this.messageRepository.save(updatedMessage);
-            return true;
+        if (optionalMessage.isEmpty()) {
+            return false;
         }
-        return false;
+        Message updatedMessage = message;
+        updatedMessage.setMessageId(message_id);
+        this.messageRepository.save(updatedMessage);
+        return true;
     }
 
-    public void deleteMessage(int message_id) {
+    public boolean deleteMessage(int message_id) {
+        if (!this.messageRepository.existsById(message_id)) {
+            return false;
+        }
         this.messageRepository.deleteById(message_id);
+        return true;
     }
 
     public List<Message> getMessagesByAccountID(int account_id) {
